+++ src/index.js
@@ -2,3 +2,4 @@ export { Builder } from "./builder";
 export * from "./convert";
 export * from "./insert";
 export * from "./replace";
+export * from "./clear";