+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2023-08-21
+
+### Added
+
+- `Builder.push` method
+- `clear` function
+
 ## [0.5.0] - 2022-09-25
 
 ### Changed
@@ -60,7 +67,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `onClick` handler
 - Wrap all builded elements by provided DOM node
 
-[Unreleased]: https://github.com/vikian050194/fandom/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/vikian050194/fandom/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/vikian050194/fandom/compare/v0.5.0...v0.6.0
+[0.5.0]: https://github.com/vikian050194/fandom/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/vikian050194/fandom/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/vikian050194/fandom/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/vikian050194/fandom/compare/v0.1.1...v0.2.0